+++ pcsx2/SPU2/defs.h
@@ -437,7 +437,7 @@ struct V_Core
 	u32 NoiseOut; // Noise Output
 	u16 AutoDMACtrl; // AutoDMA Status
 	s32 DMAICounter; // DMA Interrupt Counter
-	u32 LastClock; // DMA Interrupt Clock Cycle Counter
+	u64 LastClock; // DMA Interrupt Clock Cycle Counter
 	u32 InputDataLeft; // Input Buffer
 	u32 InputDataTransferred; // Used for simulating MADR increase (GTA VC)
 	u32 InputPosWrite;
